Refactor languages to a new directory.
