--- mocks/translation_fr.json
@@ -1,58 +1,0 @@
-{
-	"menu": {
-		"filter": "Filtre",
-		"search": "Chercher",
-		"overview": "Aperçu",
-		"api-servers": "Serveur API",
-		"authentication": "Authentification",
-		"operations": "OPÉRATIONS",
-		"components": "COMPOSANTS",
-		"schemas": "Schémas"
-	},
-	"headers": {
-		"api-servers": "SERVEUR API",
-		"authentication": "AUTHENTIFICATION"
-	},
-	"overview":{
-		"email": "Courriel",
-		"terms-of-service": "Avis"
-	},	
-	"api-servers":{
-		"server-variables": "VARIABLES SERVEUR",
-		"selected": "CHOISI"
-	},
-	"authentication":{
-		"no-api-key-applied": "Aucune clé d'API appliquée",
-		"http-basic": "HTTP Basique",
-		"http-basic-desc":"Envoyez l'en-tête Authorization contenant le type Basic suivi d'un espace et d'une chaîne encodée en base64 de nom d'utilisateur:mot de passe",
-		"username":"nom d'utilisateur",
-		"password":"mot de passe",
-		"requires": "Nécessite",
-		"http-basic-note": "un nom d'utilisateur/mot de passe encodé en base64",
-		"in-auth-header": "dans l'en-tête Authorization",
-		"set": "DÉFINIR"
-
-	},	
-	"operations": {
-		"request": "REQUÊTE", 
-		"request-body": "CORPS DE LA REQUÊTE",
-		"model": "MODÈLE",
-		"body": "CORPS",
-		"request-headers": "EN-TÊTES DE LA REQUÊTE",
-		"clear": "VIDER",
-		"clear-response": "VIDER LA RÉPONSE",
-		"execute": "EXÉCUTER",
-		"response": "RÉPONSE",
-		"response-headers": "EN-TÊTES DE LA RÉPONSE",
-		"example": "EXEMPLE",
-		"response-status": "Statut de réponse",
-		"fetch-fail": "Échec d'obtenir (Consultez l'onglet Réseau de navigateur pour plus d'information.)",
-		"copy": "Copier",
-		"copied": "Copié"
-	},
-	"schemas": {
-		"collapse-desc": "Réduire",
-		"expand-desc": "Agrandir",
-		"schema-missing": "Schéma introuvable"
-	}
-}
