+++ src/languages/en.json
@@ -1,0 +1,57 @@
+{
+  "menu": {
+    "filter": "Filter",
+    "search": "Search",
+    "overview": "Overview",
+    "api-servers": "API Servers",
+    "authentication": "Authentication",
+    "operations": "OPERATIONS",
+    "components": "COMPONENTS",
+    "schemas": "Schemas"
+  },
+  "headers": {
+    "api-servers": "API SERVER",
+    "authentication": "AUTHENTICATION"
+  },
+  "overview": {
+    "email": "Email",
+    "terms-of-service": "Terms of Service"
+  },
+  "api-servers": {
+    "server-variables": "SERVER VARIABLES",
+    "selected": "SELECTED"
+  },
+  "authentication": {
+    "no-api-key-applied": "No API key applied",
+    "http-basic": "HTTP Basic",
+    "http-basic-desc": "Send the Authorization header containing the type Basic followed by a space and a base64 encoded string of username:password",
+    "username": "username",
+    "password": "password",
+    "requires": "Requires",
+    "http-basic-note": "Base 64 encoded username:password",
+    "in-auth-header": "in Authorization header",
+    "set": "SET"
+  },
+  "operations": {
+    "request": "REQUEST",
+    "request-body": "REQUEST BODY",
+    "model": "MODEL",
+    "body": "BODY",
+    "request-headers": "REQUEST HEADERS",
+    "clear": "CLEAR",
+    "clear-response": "CLEAR RESPONSE",
+    "execute": "EXECUTE",
+    "response": "RESPONSE",
+    "response-headers": "RESPONSE HEADERS",
+    "example": "EXAMPLE",
+    "response-status": "Response Status",
+    "fetch-fail": "Failed to fetch (Check the browser network tab for more information.)",
+    "copy": "Copy",
+    "copied": "Copied"
+  },
+  "common": {
+    "collapse-desc": "Collapse",
+    "expand-desc": "Expand",
+    "schema-missing": "Schema not found"
+  }
+}
