+++ src/templates/overview-template.js
@@ -1,7 +1,7 @@
 import { html } from 'lit-element';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import { marked } from 'marked';
-import { getI18nText } from '../utils/common-utils';
+import { getI18nText } from '../languages';
 
 /* eslint-disable indent */
 export default function overviewTemplate() {
